main: move pomerium main code to an internal cmd package so that it can be called directly from tests
